Inline the drawer back into document_idle.ts
Extracting initBESDrawer to components/besDrawer.ts was cosmetic -- it
made the two sub-mains similar in length, which was a goal I invented
rather than one the split needed. The name also implied the drawer was
separable in a way it isn't: document_idle.ts initialises download,
label view, player, audio features, cart, hide/unhide and feed
alongside it, so singling out "drawer" misrepresents the file.

document_idle.ts is back to the layout main.ts had, and its test goes
back to importing it directly, which means restoring the mocks that
keep documentIdle() from running on import.
